Solution to Get nth Bit of a Number

+++ Get nth Bit of a Number@@67882a731b4392d8fc1577ea/code.c
@@ -3,7 +3,7 @@
 int main(){
     int a,b,and = 0;
     scanf("%d %d",&a,&b);
-    int c = pow(2,b);
+    int c = (int)pow(2,b);
     and = a & c;
     if(and == 0){
         printf("0");
